app/db/collections.py
from motor.motor_asyncio import AsyncIOMotorClient
from dotenv import load_dotenv
import os

load_dotenv()

MONGO_URL = os.getenv("MONGO_URL", "mongodb://localhost:27017")
client = AsyncIOMotorClient(MONGO_URL)

MONGO_DB = os.getenv("MONGO_DB", "mydatabase")
db = client[MONGO_DB]
